requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
cycler==0.11.0
Cython==0.29.21
decorator==4.4.2
docker-pycreds==0.4.0
filelock==3.9.0
gitdb==4.0.11
GitPython==3.1.40
glcontext==2.4.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.3.1
moderngl==5.8.2
mpmath==1.2.1
multipledispatch==1.0.0
networkx==2.5.1
numpy==1.19.1
Pillow==7.2.0
plyfile==1.0.1
protobuf==4.25.1
psutil==5.9.7
pyparsing==3.1.1
pyrr==0.10.3
python-dateutil==2.8.2
pytorch-fast-transformers==0.4.0
PyYAML==5.3.1
requests==2.28.1
scipy==1.5.2
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.1
simple-3dviz==0.4.0
six==1.16.0
smmap==5.0.1
sympy==1.11.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.50.2
trimesh==3.8.10
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.16.1
wxPython == 4.1.1
